+++ OneBlink.SDK/models/FormsAppUser.cs
@@ -7,15 +7,48 @@ namespace OneBlink.SDK.Model
 {
     public class FormsAppUser
     {
-        public long id { get; set; }
-        public Boolean generatePassword {get;set;}
-        public dynamic welcomeEmailParamters {get;set;}
+        public long id
+        {
+            get; set;
+        }
+        public Boolean generatePassword
+        {
+            get; set;
+        }
+        public dynamic welcomeEmailParamters
+        {
+            get; set;
+        }
         [JsonProperty]
-        public DateTime createdAt { get; internal set; }
+        public DateTime createdAt
+        {
+            get; internal set;
+        }
         [JsonProperty]
-        public DateTime updatedAt { get; internal set; }
-        public string email {get;set;}
-        public long formsAppId {get;set;}
-        public List<string> groups {get;set;}
+        public DateTime updatedAt
+        {
+            get; internal set;
+        }
+        public string email
+        {
+            get; set;
+        }
+        public long formsAppId
+        {
+            get; set;
+        }
+        public List<string> groups
+        {
+            get; set;
+        }
+        public string firstName
+        {
+            get; set;
+        }
+        public string lastName
+        {
+            get; set;
+        }
+
     }
 }