+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 -   `searchQuerystringParameter` to `FormElement` class
 -   `personalisation` and `externalIdGenerationOnSubmit` to `Form` class
 -   `FormPersonalisation` model class for `personalisation`
+-   `enableAppUserSignup` to `FormsAppBase` class
+-   `firstName` and `lastName` to `FormsAppUser`
 
 ## 4.1.8 (2023-05-12)